+++ src/main/java/com/nguyenquyhy/discordbridge/models/ChannelMinecraftConfig.java
@@ -10,26 +10,18 @@ import java.util.List;
  * Created by Hy on 10/13/2016.
  */
 @ConfigSerializable
-public class ChannelMinecraftConfig {
+public class ChannelMinecraftConfig extends ChannelMinecraftConfigCore {
+    public ChannelMinecraftConfig() {
+        initializeDefault();
+    }
+
+    @Override
     void initializeDefault() {
-        chatTemplate = "&7<%a> &f%s";
-        attachmentTemplate = "[Attachment]";
-        attachmentColor = "&3";
-        attachmentHoverTemplate = "Click to open attachment.";
-        defaultRole = "Member";
-        roleBlacklist = new ArrayList<>();
+        super.initializeDefault();
+        roles = new ArrayList<>();
     }
 
     @Setting
-    public String chatTemplate;
-    @Setting
-    public String attachmentTemplate;
-    @Setting
-    public String attachmentColor;
-    @Setting
-    public String attachmentHoverTemplate;
-    @Setting
-    public String defaultRole;
-    @Setting
-    public List<String> roleBlacklist;
+    public List<ChannelMinecraftConfigCore> roles;
+
 }