+++ src/main/java/com/nguyenquyhy/discordbridge/models/GlobalConfig.java
@@ -17,7 +17,6 @@ public class GlobalConfig {
         prefixBlacklist = new ArrayList<>();
         ignoreBots = false;
         botDiscordGame = "";
-        linkDiscordAttachments = true;
         minecraftBroadcastTemplate = "&2<BROADCAST> %s";
         botToken = "";
     }
@@ -30,8 +29,6 @@ public class GlobalConfig {
     public List<String> prefixBlacklist;
     @Setting
     public Boolean ignoreBots;
-    @Setting
-    public Boolean linkDiscordAttachments;
     @Setting
     public String botDiscordGame;
     @Setting